+++ temp.h
@@ -2,5 +2,5 @@
 #define TEMP_H
 void temp_init(void);
 void temp_run(void);
-int8_t get_temp(int16_t *temp);
+int8_t get_temp(double *temp);
 #endif
